requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
channels>=4.0
channels-redis
